blacklist: normalize phone numbers to E164 standard
Normalize to standard E164 with country support should perform better phone number matches.
xe: The brazilian user (country code +55) saves 6792928089 as blacklisted phone, but when
he receives a call from +556792928089 the blacklist provider doesn't match the number,
but both are the same number. Using E164, combined with SimCountryIso or LocaleCountryIso,
6792928089 will be stored as +556792928089, and during quering the provider the same format
will be applied for the incoming intent phone number, which will lead to a proper match.

Change-Id: I80c8330bfe3a13dca27557f795ee04781248f1a1
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>